requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
python-dotenv
requests
Werkzeug==2.0.0
itsdangerous==2.1.2
click==8.0.4
MarkupSafe==2.1.1
jinja2==3.0.0